continued root test
